feat(Options): update method call in createTicket
Updated the method call in the createTicket function in Options.php
to use the execOptions method instead of exec. This change ensures
that the correct method is called for obtaining the forms
authentication ticket.

fix(Utils/Curl): disable SSL verification

Added the CURLOPT_SSL_VERIFYPEER option with a value of false to
the curl configuration in Curl.php. This change disables SSL
verification when making POST requests, allowing for
communication with servers that have invalid or self-signed
certificates.

Issue: main